# api/openapi.yaml
openapi: 3.0.3
info:
  title: Media Manager API
  version: 0.0.0
  description: >
    Contract-first API for the self-hosted media manager. The project name is
    not final; do not use the codename in generated public identifiers.
servers:
  - url: /api
    description: Same-origin API served by the Go backend
...
paths:
...
  /media/collections/{provider}/{collectionId}:
    parameters:
      - name: provider
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/MetadataProviderType"
      - name: collectionId
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 100
    get:
      operationId: getMediaCollection
      tags:
        - Media
      summary: Get metadata for a movie collection
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Collection metadata and movies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaCollection"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /settings/custom-formats:
    get:
      operationId: listCustomFormats
      tags:
        - Settings
      summary: List custom formats
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Custom formats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomFormatListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createCustomFormat
      tags:
        - Settings
      summary: Create a custom format
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomFormatRequest"
      responses:
        "201":
          description: Custom format created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomFormat"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /settings/custom-formats/test-parsing:
    post:
      operationId: testCustomFormatParsing
      tags:
        - Settings
      summary: Test custom format parsing for a release title
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomFormatParsingRequest"
      responses:
        "200":
          description: Parsed release details and matched custom formats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomFormatParsingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /settings/custom-formats/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    put:
      operationId: updateCustomFormat
      tags:
        - Settings
      summary: Update a custom format
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomFormatRequest"
      responses:
        "200":
          description: Custom format updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomFormat"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteCustomFormat
      tags:
        - Settings
      summary: Delete a custom format
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Custom format deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /settings/library/folder-options:
    get:
      operationId: listLibraryFolderOptions
      tags:
        - Settings
      summary: Browse server-visible folders for library selection
      security:
        - sessionCookie: []
      parameters:
        - name: path
          in: query
          schema:
            type: string
            maxLength: 4000
      responses:
        "200":
          description: Folder picker options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryFolderOptionListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createLibraryFolderOption
      tags:
        - Settings
      summary: Create a server-visible child folder for library selection
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryFolderOptionCreateRequest"
      responses:
        "201":
          description: Folder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryFolderOption"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /settings/library/folders:
    get:
      operationId: listLibraryFolders
      tags:
        - Settings
      summary: List configured library folders
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Configured library folders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryFolderListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createLibraryFolder
      tags:
        - Settings
      summary: Add a library folder and scan it for media
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryFolderRequest"
      responses:
        "201":
          description: Library folder added and scan completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryFolderCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /settings/library/folders/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    delete:
      operationId: deleteLibraryFolder
      tags:
        - Settings
      summary: Remove a configured library folder
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Library folder removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /settings/library/folders/{id}/scan:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      operationId: scanLibraryFolder
      tags:
        - Settings
      summary: Scan an existing library folder for media
      security:
        - sessionCookie: []
      responses:
        "201":
          description: Library folder scan completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryScan"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /settings/library/scans/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      operationId: getLibraryScan
      tags:
        - Settings
      summary: Get a library folder scan with discovered media review items
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Library scan review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryScan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /settings/library/scans/{id}/items/{itemId}/match:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
      - name: itemId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: matchLibraryScanItem
      tags:
        - Settings
      summary: Manually match a discovered file to a movie, series, or anime item
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryScanItemMatchRequest"
      responses:
        "200":
          description: Discovered item matched and added to monitored media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryScanItemMatchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or session invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    MediaType:
      type: string
      enum:
        - movie
        - series
...
    MediaSearchResult:
      type: object
      additionalProperties: false
      required:
        - title
        - type
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        type:
          $ref: "#/components/schemas/MediaType"
        year:
          type: integer
          format: int32
        externalProvider:
          type: string
        externalId:
          type: string
        overview:
          type: string
        posterPath:
          type: string
        collectionId:
          type: string
        collectionName:
          type: string
...
    MediaCollection:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - provider
        - results
      properties:
        id:
          type: string
        name:
          type: string
        provider:
          $ref: "#/components/schemas/MetadataProviderType"
        overview:
          type: string
        posterPath:
          type: string
        backdropPath:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/MediaSearchResult"
...
    MediaMonitorMode:
      type: string
      enum:
        - only_media
        - collection
...
    CustomFormatListResponse:
      type: object
      additionalProperties: false
      required:
        - formats
      properties:
        formats:
          type: array
          items:
            $ref: "#/components/schemas/CustomFormat"
    CustomFormat:
      allOf:
        - $ref: "#/components/schemas/CustomFormatRequest"
        - type: object
          additionalProperties: false
          required:
            - id
            - createdAt
            - updatedAt
          properties:
            id:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    CustomFormatRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - includeSpecs
        - excludeSpecs
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        includeSpecs:
          type: array
          items:
            $ref: "#/components/schemas/CustomFormatSpec"
        excludeSpecs:
          type: array
          items:
            $ref: "#/components/schemas/CustomFormatSpec"
    CustomFormatSpec:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - type
        - value
        - required
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 80
        name:
          type: string
          minLength: 1
          maxLength: 120
        type:
          $ref: "#/components/schemas/CustomFormatSpecType"
        value:
          type: string
          minLength: 1
          maxLength: 500
        required:
          type: boolean
...
    CustomFormatParsingRequest:
      type: object
      additionalProperties: false
      required:
        - fileName
      properties:
        fileName:
          type: string
          minLength: 1
          maxLength: 500
    CustomFormatParsingResponse:
      type: object
      additionalProperties: false
      required:
        - fileName
        - release
        - quality
        - languages
        - details
        - calculatedScore
        - matchedCustomFormats
      properties:
        fileName:
          type: string
        release:
          $ref: "#/components/schemas/ParsedReleaseInfo"
        quality:
          $ref: "#/components/schemas/ParsedQualityInfo"
        languages:
          type: array
          items:
            type: string
        details:
          $ref: "#/components/schemas/ParsedReleaseDetails"
        matchedProfile:
          $ref: "#/components/schemas/CustomFormatParsingProfile"
        calculatedScore:
          type: integer
          format: int32
        matchedCustomFormats:
          type: array
          items:
            $ref: "#/components/schemas/CustomFormatParsingMatch"
    CustomFormatParsingProfile:
      type: object
      additionalProperties: false
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
    ParsedReleaseInfo:
      type: object
      additionalProperties: false
      required:
        - releaseTitle
        - movieTitle
        - year
        - edition
        - releaseGroup
        - releaseHash
      properties:
        releaseTitle:
          type: string
        movieTitle:
          type: string
        year:
          type: string
        edition:
          type: string
        releaseGroup:
          type: string
        releaseHash:
          type: string
    ParsedQualityInfo:
      type: object
      additionalProperties: false
      required:
        - qualityId
        - quality
        - source
        - resolution
        - videoCodec
        - audioCodec
        - audioChannels
        - version
        - proper
        - repack
        - real
      properties:
        qualityId:
          type: string
        quality:
          type: string
        source:
          type: string
        resolution:
          type: string
        videoCodec:
          type: string
        audioCodec:
          type: string
        audioChannels:
          type: string
        version:
          type: string
        proper:
          type: boolean
        repack:
          type: boolean
        real:
          type: boolean
    ParsedReleaseDetails:
      type: object
      additionalProperties: false
      required:
        - releaseType
        - customFormatNames
        - matchedSpecCount
      properties:
        releaseType:
          type: string
        customFormatNames:
          type: array
          items:
            type: string
        matchedSpecCount:
          type: integer
          format: int32
    CustomFormatParsingMatch:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - score
        - matchedSpecs
      properties:
        id:
          type: string
        name:
          type: string
        score:
          type: integer
          format: int32
        matchedSpecs:
          type: array
          items:
            $ref: "#/components/schemas/CustomFormatSpec"
...
    MetadataProviderType:
      type: string
      enum:
        - tmdb
        - tvdb
...
    LibraryFolderListResponse:
      type: object
      additionalProperties: false
      required:
        - folders
      properties:
        folders:
          type: array
          items:
            $ref: "#/components/schemas/LibraryFolder"
    LibraryFolderCreateResponse:
      type: object
      additionalProperties: false
      required:
        - folder
        - scan
      properties:
        folder:
          $ref: "#/components/schemas/LibraryFolder"
        scan:
          $ref: "#/components/schemas/LibraryScan"
    LibraryFolder:
      allOf:
        - $ref: "#/components/schemas/LibraryFolderRequest"
        - type: object
          additionalProperties: false
          required:
            - id
            - createdAt
            - updatedAt
          properties:
            id:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    LibraryFolderRequest:
      type: object
      additionalProperties: false
      required:
        - path
      properties:
        path:
          type: string
          minLength: 1
          maxLength: 4000
...
    LibraryFolderOptionListResponse:
      type: object
      additionalProperties: false
      required:
        - currentPath
        - entries
      properties:
        currentPath:
          type: string
        parentPath:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LibraryFolderOption"
    LibraryFolderOption:
      type: object
      additionalProperties: false
      required:
        - name
        - path
      properties:
        name:
          type: string
        path:
          type: string
    LibraryFolderOptionCreateRequest:
      type: object
      additionalProperties: false
      required:
        - parentPath
        - name
      properties:
        parentPath:
          type: string
          minLength: 1
          maxLength: 4000
        name:
          type: string
          minLength: 1
          maxLength: 255
    LibraryMediaKind:
      type: string
      enum:
        - movie
        - series
        - anime_movie
        - anime_series
        - unknown
...
    LibraryScan:
      type: object
      additionalProperties: false
      required:
        - id
        - folderId
        - folderPath
        - status
        - totalFiles
        - autoMatchedCount
        - manualCount
        - items
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        folderId:
          type: string
          format: uuid
        folderPath:
          type: string
        status:
          type: string
          enum:
            - completed
            - failed
        totalFiles:
          type: integer
          format: int32
          minimum: 0
        autoMatchedCount:
          type: integer
          format: int32
          minimum: 0
        manualCount:
          type: integer
          format: int32
          minimum: 0
        items:
          type: array
          items:
            $ref: "#/components/schemas/LibraryScanItem"
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    LibraryScanItem:
      type: object
      additionalProperties: false
      required:
        - id
        - scanId
        - path
        - fileName
        - detectedTitle
        - detectedMediaKind
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        scanId:
          type: string
          format: uuid
        path:
          type: string
        fileName:
          type: string
        detectedTitle:
          type: string
        detectedYear:
          type: integer
          format: int32
          minimum: 1800
          maximum: 3000
        detectedMediaKind:
          $ref: "#/components/schemas/LibraryMediaKind"
        status:
          $ref: "#/components/schemas/LibraryScanItemStatus"
        matchedTitle:
          type: string
        matchedYear:
          type: integer
          format: int32
          minimum: 1800
          maximum: 3000
        matchedMediaKind:
          $ref: "#/components/schemas/LibraryMediaKind"
        mediaItemId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LibraryScanItemMatchRequest:
      type: object
      additionalProperties: false
      required:
        - mediaKind
        - title
        - monitored
        - qualityProfileId
        - monitorMode
        - minimumAvailability
      properties:
        mediaKind:
          $ref: "#/components/schemas/LibraryMediaKind"
        title:
          type: string
          minLength: 1
          maxLength: 500
        year:
          type: integer
          format: int32
          minimum: 1800
          maximum: 3000
        monitored:
          type: boolean
        qualityProfileId:
          type: string
          minLength: 1
          maxLength: 100
        monitorMode:
          $ref: "#/components/schemas/MediaMonitorMode"
        minimumAvailability:
          $ref: "#/components/schemas/MinimumAvailability"
        externalProvider:
          type: string
          maxLength: 100
        externalId:
          type: string
          maxLength: 200
        overview:
          type: string
          maxLength: 5000
        posterPath:
          type: string
          maxLength: 2000
    LibraryScanItemMatchResponse:
      type: object
      additionalProperties: false
      required:
        - item
        - mediaItem
      properties:
        item:
          $ref: "#/components/schemas/LibraryScanItem"
        mediaItem:
          $ref: "#/components/schemas/MediaItem"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true